requirements.txt
fastapi
uvicorn
pillow
python-multipart
pytesseract
